backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Request parsing
python-multipart==0.0.9

# Environment variables
python-dotenv==1.0.1

# Data validation
pydantic==2.7.1

# LLM
groq==0.9.0
httpx==0.27.2

# RAG
sentence-transformers==2.7.0
supabase==2.6.0

# Optional image endpoint support in the existing codebase
google-generativeai==0.8.5
pillow==10.3.0
